+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Hasura GraphQL Engine Changelog
 
 ## Next release
+
+
+### Bug fixes and improvements
+
 (Add entries below in the order of: server, console, cli, docs, others)
 
+- server: fix regression: `on_conflict` was missing in the schema for inserts in tables where the current user has no columns listed in their update permissions (fix #6804)
 - cli: fix regression - `metadata apply —dry-run` was overwriting local metadata files with metadata on server when it should just display the differences.
 - cli: add support for `api_limits` metadata object